moved isapprox and scalar to multivectors.jl
