+++ src/multivectors.jl
@@ -311,8 +311,8 @@ end
 
 ## Generic
 
-import Base: isinf
-export basis, grade, hasinf, hasorigin, isorigin
+import Base: isinf, isapprox
+export basis, grade, hasinf, hasorigin, isorigin, scalar
 
 const VBV = Union{MValue,SValue,MBlade,SBlade,MultiVector}
 
@@ -335,6 +335,23 @@ const VBV = Union{MValue,SValue,MBlade,SBlade,MultiVector}
 @pure hasorigin(e::Basis{V}) where V = hasorigin(V) && (hasinf(V) ? e[2] : isodd(bits(e)))
 @pure hasorigin(t::Union{MValue,SValue}) = hasorigin(basis(t))
 @pure hasorigin(m::TensorAlgebra) = hasorigin(vectorspace(m))
+
+function isapprox(a::TensorMixed{T1}, b::TensorMixed{T2}) where {T1, T2}
+    rtol = Base.rtoldefault(T1, T2, 0)    
+    return norm(value(a-b)) <= rtol * max(norm(value(a)), norm(value(b)))
+end
+isapprox(a::TensorMixed, b::TensorTerm) = isapprox(a, MultiVector(b))
+isapprox(b::TensorTerm, a::TensorMixed) = isapprox(a, MultiVector(b))
+isapprox(a::TensorTerm, b::TensorTerm) = isapprox(MultiVector(a), MultiVector(b))
+
+"""
+    scalar(multivector)
+    
+Return the scalar (grade 0) part of any multivector.
+"""
+scalar(a::TensorMixed) = grade(a) == 0 ? a[1] : 0
+scalar(a::MultiVector) = a[0][1]
+scalar(a::TensorAlgebra) = scalar(MultiVector(a))
 
 ## MultiGrade{N}
